use texture object

+++ src/lprec/solver.py
@@ -16,8 +16,7 @@ pinned_memory_pool = cp.cuda.PinnedMemoryPool()
 cp.cuda.set_pinned_memory_allocator(pinned_memory_pool.malloc)
 
 # 'float32'- c++ code needs to be recompiled with changed directives in cfunc.cuh
-mtype = 'float32'
-
+mtype = 'float16'
 
 class LpRec(cfunc):
